pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bt_api_ib_web"
version = "2.0.0"
description = "Interactive Brokers Web API plugin for bt_api"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bt_api_base>=0.15,<1.0",
    "pydantic>=2",
    "PyYAML>=6",
    "python-dotenv>=1",
    "requests>=2.31",
    "websocket-client>=1.7",
]

[project.entry-points."bt_api.plugins"]
ib_web = "bt_api_ib_web.plugin:register_plugin"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bt_api_ib_web = ["configs/*.yaml"]

[tool.ruff]
src = ["src"]
fix = true
show-fixes = false

[tool.ruff.lint]
extend-select = ["B", "E", "F", "I", "Q", "RUF", "UP", "W"]
ignore = ["E501", "RUF012", "Q000", "Q003"]

[tool.ruff.format]
quote-style = "single"
